refactor: move to shared
